pyproject.toml
[tool.poetry]
name = "promodoro-poetry"
version = "0.1.0"
description = ""
authors = ["Danilov Aleksandr <danilov.aa1@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic-settings = "^2.6.1"
orjson = "^3.10.12"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.2.1"
python-multipart = "^0.0.19"
pytest = {extras = ["async"], version = "^8.3.4"}
pytest-asyncio = "^0.25.0"
pytest-dotenv = "^0.5.2"
celery = "^5.4.0"
redis = "^5.2.1"
gevent = "^24.11.1"
gunicorn = "^23.0.0"
aio-pika = "^9.5.4"
aiokafka = "^0.12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
